airflow-data-pipeline/requirements.txt
apache-airflow==2.10.2
pandas==2.2.3